Levels where exerciser test are executed
- Exerciser table is created with all PCIe devices,
  following exerciser tests can be moved to BSA and
  removed from SBSA.
    SBSA e002: move to BSA test_os_e016
    SBSA e004: move to BSA test_os_e017
    SBSA e005: move to BSA test_os_e018
- rename of other exerciser test
