+++ test_pool/exerciser/operating_system/test_e003.c
@@ -28,10 +28,10 @@
 #include "val/include/sbsa_avs_exerciser.h"
 
 #define TEST_NUM   (AVS_EXERCISER_TEST_NUM_BASE + 3)
-#define TEST_DESC  "PCIe Address translation check    "
+#define TEST_DESC  "ATS Functionality Check           "
 #define TEST_RULE  "RE_SMU_2"
 
-#define TEST_DATA_NUM_PAGES  4
+#define TEST_DATA_NUM_PAGES  1
 #define TEST_DATA 0xDE
 
 static
@@ -70,12 +70,14 @@ payload(void)
   uint32_t dma_len;
   uint32_t instance;
   uint32_t e_bdf;
+  uint32_t cap_base;
   void *dram_buf_in_virt;
   void *dram_buf_out_virt;
   uint64_t dram_buf_in_phys;
   uint64_t dram_buf_out_phys;
   uint64_t dram_buf_in_iova;
   uint64_t dram_buf_out_iova;
+  uint64_t m_vir_addr;
   uint32_t num_exercisers, num_smmus;
   uint32_t device_id, its_id;
   uint32_t page_size = val_memory_page_size();
@@ -85,24 +87,27 @@ payload(void)
   uint64_t ttbr;
   uint32_t test_data_blk_size = page_size * TEST_DATA_NUM_PAGES;
   uint64_t *pgt_base_array;
-
-  num_smmus = val_iovirt_get_smmu_info(SMMU_NUM_CTRL, 0);
+  uint64_t translated_addr;
+  uint32_t test_skip = 1;
+  uint32_t reg_value = 0;
 
   /* Initialize DMA master and memory descriptors */
   val_memory_set(&master, sizeof(master), 0);
   val_memory_set(mem_desc_array, sizeof(mem_desc_array), 0);
   mem_desc = &mem_desc_array[0];
   dram_buf_in_phys = 0;
+
   pe_index = val_pe_get_index_mpid(val_pe_get_mpid());
   num_exercisers = val_exerciser_get_info(EXERCISER_NUM_CARDS, 0);
+  num_smmus = val_iovirt_get_smmu_info(SMMU_NUM_CTRL, 0);
 
   /* Allocate an array to store base addresses of page tables allocated for
    * all exercisers
    */
   pgt_base_array = val_aligned_alloc(MEM_ALIGN_4K, sizeof(uint64_t) * num_exercisers);
   if (!pgt_base_array) {
-      val_print(AVS_PRINT_ERR, "\n       mem alloc failure %x", 03);
-      val_set_status(pe_index, RESULT_FAIL(g_sbsa_level, TEST_NUM, 03));
+      val_print(AVS_PRINT_ERR, "\n       mem alloc failure", 0);
+      val_set_status(pe_index, RESULT_FAIL(g_sbsa_level, TEST_NUM, 02));
       return;
   }
 
@@ -111,9 +116,9 @@ payload(void)
   /* Allocate a buffer to perform DMA tests on */
   dram_buf_in_virt = val_memory_alloc_pages(TEST_DATA_NUM_PAGES);
   if (!dram_buf_in_virt) {
-      val_print(AVS_PRINT_ERR, "\n       Cacheable mem alloc failure %x", 02);
+      val_print(AVS_PRINT_ERR, "\n       Cacheable mem alloc failure", 0);
       val_memory_free_aligned(pgt_base_array);
-      val_set_status(pe_index, RESULT_FAIL(g_sbsa_level, TEST_NUM, 02));
+      val_set_status(pe_index, RESULT_FAIL(g_sbsa_level, TEST_NUM, 03));
       return;
   }
 
@@ -124,25 +129,15 @@ payload(void)
   dma_len = test_data_blk_size / 2;
 
   /* Get translation attributes via TCR and translation table base via TTBR */
-  if (val_pe_reg_read_tcr(0 /*for TTBR0*/, &pgt_desc.tcr)) {
-    val_print(AVS_PRINT_ERR, "\n       Unable to get translation attributes via TCR", 0);
+  if (val_pe_reg_read_tcr(0 /*for TTBR0*/,
+                          &pgt_desc.tcr)) {
+    val_print(AVS_PRINT_ERR, "\n       TCR read failure", 0);
     goto test_fail;
   }
 
-  if (val_pe_reg_read_ttbr(0 /*TTBR0*/, &ttbr)) {
-    val_print(AVS_PRINT_ERR, "\n       Unable to get translation table via TBBR", 0);
-    goto test_fail;
-  }
-
-  pgt_desc.pgt_base = (ttbr & AARCH64_TTBR_ADDR_MASK);
-  pgt_desc.mair = val_pe_reg_read(MAIR_ELx);
-  pgt_desc.stage = PGT_STAGE1;
-
-  /* Get memory attributes of the test buffer, we'll use the same attibutes to create
-   * our own page table later.
-   */
-  if (val_pgt_get_attributes(pgt_desc, (uint64_t)dram_buf_in_virt, &mem_desc->attributes)) {
-    val_print(AVS_PRINT_ERR, "\n       Unable to get memory attributes of the test buffer", 0);
+  if (val_pe_reg_read_ttbr(0 /*for TTBR0*/,
+                           &ttbr)) {
+    val_print(AVS_PRINT_ERR, "\n       TTBR0 read failure", 0);
     goto test_fail;
   }
 
@@ -158,7 +153,27 @@ payload(void)
 
     /* Get exerciser bdf */
     e_bdf = val_exerciser_get_bdf(instance);
-    val_print(AVS_PRINT_DEBUG, "\n       Exercise BDF - 0x%x", e_bdf);
+    val_print(AVS_PRINT_DEBUG, "\n       Exerciser BDF - 0x%x", e_bdf);
+
+    /* If ATS Capability Not Present, Skip. */
+    if (val_pcie_find_capability(e_bdf, PCIE_ECAP, ECID_ATS, &cap_base) != PCIE_SUCCESS)
+        continue;
+
+    val_pcie_read_cfg(e_bdf, cap_base + ATS_CTRL, &reg_value);
+    reg_value |= ATS_CACHING_EN;
+    val_pcie_write_cfg(e_bdf, cap_base + ATS_CTRL, reg_value);
+
+    pgt_desc.pgt_base = (ttbr & AARCH64_TTBR_ADDR_MASK);
+    pgt_desc.mair = val_pe_reg_read(MAIR_ELx);
+    pgt_desc.stage = PGT_STAGE1;
+
+    /* Get memory attributes of the test buffer, we'll use the same attibutes to create
+     * our own page table later.
+     */
+    if (val_pgt_get_attributes(pgt_desc, (uint64_t)dram_buf_in_virt, &mem_desc->attributes)) {
+        val_print(AVS_PRINT_ERR, "\n       Unable to get memory attributes of the test buffer", 0);
+        goto test_fail;
+    }
 
     /* Get SMMU node index for this exerciser instance */
     master.smmu_index = val_iovirt_get_rc_smmu_index(PCIE_EXTRACT_BDF_SEG(e_bdf), PCIE_CREATE_BDF_PACKED(e_bdf));
@@ -187,26 +202,26 @@ payload(void)
 
         /* Need to know input and output address sizes before creating page table */
         pgt_desc.ias = val_smmu_get_info(SMMU_IN_ADDR_SIZE, master.smmu_index);
-        if (pgt_desc.ias == 0) {
-          val_print(AVS_PRINT_ERR,
-                    "\n       Input address size of SMMU %d is 0", master.smmu_index);
-          goto test_fail;
+        if ((pgt_desc.ias) == 0) {
+            val_print(AVS_PRINT_ERR,
+                          "\n       Input address size of SMMU %d is 0", master.smmu_index);
+            goto test_fail;
         }
 
         pgt_desc.oas = val_smmu_get_info(SMMU_OUT_ADDR_SIZE, master.smmu_index);
-        if (pgt_desc.oas == 0) {
-          val_print(AVS_PRINT_ERR,
-                    "\n       Output address size of SMMU %d is 0", master.smmu_index);
-          goto test_fail;
+        if ((pgt_desc.oas) == 0) {
+            val_print(AVS_PRINT_ERR,
+                          "\n       Output address size of SMMU %d is 0", master.smmu_index);
+            goto test_fail;
         }
 
         /* set pgt_desc.pgt_base to NULL to create new translation table, val_pgt_create
            will update pgt_desc.pgt_base to point to created translation table */
         pgt_desc.pgt_base = (uint64_t) NULL;
         if (val_pgt_create(mem_desc, &pgt_desc)) {
-          val_print(AVS_PRINT_ERR,
-                    "\n       Unable to create page table with given attributes", 0);
-          goto test_fail;
+            val_print(AVS_PRINT_ERR,
+                      "\n       Unable to create page table with given attributes", 0);
+            goto test_fail;
         }
 
         pgt_base_array[instance] = pgt_desc.pgt_base;
@@ -214,8 +229,7 @@ payload(void)
         /* Configure the SMMU tables for this exerciser to use this page table for VA to PA translations*/
         if (val_smmu_map(master, pgt_desc))
         {
-            val_print(AVS_PRINT_ERR,
-                     "\n       SMMU mapping failed (%x)     ", e_bdf);
+            val_print(AVS_PRINT_ERR, "\n       SMMU mapping failed (%x)     ", e_bdf);
             goto test_fail;
         }
 
@@ -223,10 +237,36 @@ payload(void)
         dram_buf_out_iova = dram_buf_in_iova + (test_data_blk_size / 2);
     }
 
+    test_skip = 0;
+
+    val_exerciser_set_param(DMA_ATTRIBUTES, (uint64_t)dram_buf_in_virt + instance * test_data_blk_size, dma_len, instance);
+
+    /* Send an ATS Translation Request for the VA */
+    if (val_exerciser_ops(ATS_TXN_REQ, (uint64_t)dram_buf_in_virt + instance * test_data_blk_size, instance)) {
+        val_print(AVS_PRINT_ERR, "\n       ATS Translation Req Failed exerciser %4x", instance);
+        goto test_fail;
+    }
+
+    /* Get ATS Translation Response */
+    m_vir_addr = (uint64_t)dram_buf_in_virt + instance * test_data_blk_size;
+    if (val_exerciser_get_param(ATS_RES_ATTRIBUTES, &translated_addr, &m_vir_addr, instance)) {
+        val_print(AVS_PRINT_ERR, "\n       ATS Response failure %4x", instance);
+        goto test_fail;
+    }
+
+    /* Compare Translated Addr with Physical Address from the Mappings */
+    if (translated_addr != dram_buf_in_phys) {
+        val_print(AVS_PRINT_ERR, "\n       ATS Translation failure %4x", instance);
+        goto test_fail;
+    }
+
     /* Initialize the sender buffer with test specific data */
     write_test_data(dram_buf_in_virt, dma_len);
 
-    if (val_exerciser_set_param(DMA_ATTRIBUTES, dram_buf_in_iova, dma_len, instance)) {
+    /* Configure Exerciser to issue subsequent DMA transactions with AT(Address Translated) bit Set */
+    val_exerciser_set_param(CFG_TXN_ATTRIBUTES, TXN_ADDR_TYPE, AT_TRANSLATED, instance);
+
+    if (val_exerciser_set_param(DMA_ATTRIBUTES, dram_buf_in_phys, dma_len, instance)) {
         val_print(AVS_PRINT_ERR, "\n       DMA attributes setting failure %4x", instance);
         goto test_fail;
     }
@@ -253,14 +293,20 @@ payload(void)
         goto test_fail;
     }
 
+
     clear_dram_buf(dram_buf_in_virt, test_data_blk_size);
+
   }
 
-  val_set_status(pe_index, RESULT_PASS(g_sbsa_level, TEST_NUM, 01));
+  if (test_skip)
+    val_set_status(pe_index, RESULT_SKIP(g_sbsa_level, TEST_NUM, 01));
+  else
+    val_set_status(pe_index, RESULT_PASS(g_sbsa_level, TEST_NUM, 01));
+
   goto test_clean;
 
 test_fail:
-  val_set_status(pe_index, RESULT_FAIL(g_sbsa_level, TEST_NUM, 02));
+  val_set_status(pe_index, RESULT_FAIL(g_sbsa_level, TEST_NUM, 01));
 
 test_clean:
   /* Return the pages to the heap manager */
@@ -276,11 +322,21 @@ test_clean:
                                    &device_id, &master.streamid,
                                    &its_id))
         continue;
+
     val_smmu_unmap(master);
+
     if (pgt_base_array[instance] != 0) {
       pgt_desc.pgt_base = pgt_base_array[instance];
       val_pgt_destroy(pgt_desc);
     }
+
+    if (val_pcie_find_capability(e_bdf, PCIE_ECAP, ECID_ATS, &cap_base) == PCIE_SUCCESS)
+    {
+        val_pcie_read_cfg(e_bdf, cap_base + ATS_CTRL, &reg_value);
+        reg_value &= ATS_CACHING_DIS;
+        val_pcie_write_cfg(e_bdf, cap_base + ATS_CTRL, reg_value);
+    }
+
   }
 
   /* Disable all SMMUs */
