+++ uefi_app/SbsaAvsNist.inf
@@ -141,9 +141,6 @@
   ../test_pool/exerciser/operating_system/test_e006.c
   ../test_pool/exerciser/operating_system/test_e007.c
   ../test_pool/exerciser/operating_system/test_e008.c
-  ../test_pool/exerciser/operating_system/test_e009.c
-  ../test_pool/exerciser/operating_system/test_e010.c
-  ../test_pool/exerciser/operating_system/test_e011.c
 
   ../test_pool/mpam/operating_system/test_mpam001.c
   ../test_pool/mpam/operating_system/test_mpam002.c
